refactor: rm profile and friendprofile add new prifile
