+++ frontend/index.html
@@ -15,7 +15,6 @@
 	<link rel="stylesheet" href="./src/components/Header/Header.css">
 	<link rel="stylesheet" href="./src/components/Profile/Profile.css">
 	<link rel="stylesheet" href="./src/components/Profile/FriendProfile.css">
-	<link rel="stylesheet" href="./src/components/Profile/ProfileIcon.css">
 	<link rel="stylesheet" href="./src/components/SelectPage/SelectPage.css">
 	<link rel="stylesheet" href="./src/components/Sidebar/Sidebar.css">
 	<link rel="stylesheet" href="./src/components/InfoFriendModal/InfoFriendModal.css">
