+++ frontend/src/components/Header/Header.js
@@ -1,7 +1,6 @@
 import { Component, getCookie } from "../../core/core.js";
-import { CharactorProfile } from "../Profile/Profile.js";
 import { MyProfileContent } from "../MyProfileContent/MyProfileContent.js";
-import { Profile } from "../Profile/Charactor.js";
+import { Profile } from "../Profile/Profile.js";
 
 export class Header extends Component {
     constructor(props) {
@@ -16,14 +15,12 @@ export class Header extends Component {
             <div class="profile"></div>
         `;
         const player = getCookie('player');
-        // const profile = new CharactorProfile(props.profile_img, player);
         const profile = new Profile(props.profile_img, player, 'l');
         this.el.querySelector('.profile').appendChild(profile.el);
         this.el.querySelector('.return').addEventListener('click', () => {
             location.href = '/#/';
         });
         
-        // CharactorProfile 클릭 이벤트: SelectPage의 내용을 MyProfileContent로 변경
         this.el.querySelector('.profile').addEventListener('click', () => {
             const selectPagePos = document.querySelector('.selectpagepos'); // SelectPage 위치 찾기
             selectPagePos.innerHTML = ''; // 현재 SelectPage 내용을 지움
