--- frontend/src/components/Profile/FriendProfile.css
@@ -1,46 +1,0 @@
-/* 프로필 친구 목록 스타일 */
-.charactor-profile {
-    display: flex;
-    align-items: center;
-    border: none;
-    background: none;
-    padding: 10px;
-    cursor: pointer;
-    transition: background-color 0.3s;
-}
-
-.charactor-profile .frame {
-    width: 50px;
-    height: 50px;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    background-color: white;
-    border-radius: 50%;
-    border: 3px solid gray;
-    margin-right: 10px;
-    overflow: hidden;
-}
-
-.charactor-profile img {
-    width: 100%;
-    height: 100%;
-    object-fit: cover;
-}
-
-.charactor-profile .info-container {
-    display: flex;
-    flex-direction: column; /* 정보들을 세로로 배치 */
-}
-
-.charactor-profile .info-container span {
-    font-family: 'Pixeboy', sans-serif;
-    font-size: 40px;
-    color: white;
-    background: none;
-}
-
-.charactor-profile .info-container span:nth-child(2) {
-    font-size: 15px; /* 두 번째 span의 폰트 크기를 더 작게 설정 */
-    color: lightgray; /* 두 번째 span의 폰트 색상을 변경 */
-}
