--- frontend/src/components/Profile/Profile.css
@@ -1,23 +1,0 @@
-/* 헤더 프로필 */
-.my-charactor-profile {
-    display: flex; /* 이미지와 텍스트를 가로로 배치 */
-    align-items: center; /* 세로 중앙 정렬 */
-    border: none; /* 버튼 기본 테두리 제거 */
-    background: none; /* 버튼 기본 배경 제거 */
-    padding: 10px; /* 버튼 내 여백 추가 */
-    cursor: pointer; /* 클릭할 수 있는 요소로 설정 */
-    transition: background-color 0.3s ease; /* 클릭 시 배경색 변화에 부드러운 전환 추가 */
-}
-
-.my-charactor-profile .frame {
-    display: flex; /* 이미지 중앙 정렬을 위한 Flex 설정 */
-    align-items: center; /* 세로 중앙 정렬 */
-    justify-content: center; /* 가로 중앙 정렬 */
-    margin-right: 10px; /* 이미지와 텍스트 사이의 간격 추가 */
-}
-
-.my-charactor-profile span {
-    font-family: 'Pixeboy', sans-serif; /* 폰트 설정 (예시) */
-    font-size: 20px; /* 폰트 크기 */
-    color: white; /* 폰트 색상 */
-}
