--- frontend/src/components/Profile/ProfileIcon.css
@@ -1,19 +1,0 @@
-/* 프로필 아이콘 */
-.frame_icon {
-    width: 50px; /* 프레임의 너비 */
-    height: 50px; /* 프레임의 높이 */
-    display: flex; /* 이미지 중앙 정렬을 위한 flex 설정 */
-    align-items: center; /* 세로 중앙 정렬 */
-    justify-content: center; /* 가로 중앙 정렬 */
-    background-color: white; /* 원형 테두리의 배경 색상 */
-    border-radius: 50%; /* 원형 테두리 */
-    border: 3px solid gray; /* 회색 원형 테두리 */
-    margin-right: 10px; /* 이미지와 텍스트 간의 간격 */
-    overflow: hidden; /* 이미지가 테두리를 넘지 않도록 숨김 */
-}
-
-.profile-icon{
-    width: 100%; /* 프레임에 맞게 이미지 크기 조정 */
-    height: 100%; /* 프레임에 맞게 이미지 크기 조정 */
-    object-fit: cover; /* 이미지 비율 유지하며 잘라내기 */
-}
